refactor: types 및 view 디렉토리 리팩토링
