--- src/types/lol.types.ts
@@ -1,17 +1,0 @@
-export type QueueType = 'RANKED_SOLO_5x5' | 'RANKED_FLEX_SR'
-
-export type Lane = 'TOP' | 'JUNGLE' | 'MIDDLE' | 'BOTTOM' | 'UTILITY' | 'ANY'
-
-export type Tier =
-  | 'IRON'
-  | 'BRONZE'
-  | 'SILVER'
-  | 'GOLD'
-  | 'PLATINUM'
-  | 'EMERALD'
-  | 'DIAMOND'
-  | 'MASTER'
-  | 'GRANDMASTER'
-  | 'CHALLENGER'
-
-export type RANK = 'I' | 'II' | 'III' | 'IV'
