+++ src/views/SummonerView.ts
@@ -1,12 +1,12 @@
 import { EmbedBuilder } from 'discord.js'
-import { IRankStat } from '../models/rankStat.model'
-import { ISummoner } from '../models/summoner.model'
+import { IRankStat } from '../models/RankStat'
+import { ISummoner } from '../models/Summoner'
 
 export interface SummonerViewDto {
   summoner: ISummoner
   rankStat: IRankStat
 }
 
-export default abstract class SummonerView {
+export abstract class SummonerView {
   abstract createEmbed(dto: SummonerViewDto): EmbedBuilder
 }
