--- src/views/View.ts
@@ -1,9 +1,0 @@
-import { MessagePayload, InteractionUpdateOptions } from 'discord.js'
-
-export interface ViewDto {}
-
-export default abstract class View {
-  abstract createReply(
-    dto: ViewDto,
-  ): string | MessagePayload | InteractionUpdateOptions
-}
