+++ src/views/index.ts
@@ -1,0 +1,3 @@
+export * from './BasicSummonerView'
+export * from './DetailedSummonerView'
+export * from './SummonerView'
